requirements.txt
numpy
sounddevice
keyboard
openai-whisper
torch
webrtcvad
requests
colorama
gtts
flask